Support compiling with -Werror

+++ source/platform/nspire/d_fill.c
@@ -48,9 +48,9 @@ void D_FillRect (vrect_t *rect, int color)
 		rheight += ry;
 		ry = 0;
 	}
-	if (rx+rwidth > vid.width)
+	if (rx+rwidth > (int)vid.width)
 		rwidth = vid.width - rx;
-	if (ry+rheight > vid.height)
+	if (ry+rheight > (int)vid.height)
 		rheight = vid.height - rx;
 		
 	if (rwidth < 1 || rheight < 1)
